Kansas City New Restaurant Email Notifier
This python program searches and downloads the Kansas City MO Government list of business licenses and filters for restaurants. It emails the new businesses daily via cron. Although it is intended for this city, it can easily be adapted for any city. It was created in one day, on 27 August 2025 as a demonstration for a job interview.